chore: add better documentation

+++ text-recognition/android/src/main/java/com/rnmlkit/textrecognition/TextRecognitionModule.java
@@ -101,9 +101,9 @@ public class TextRecognitionModule extends ReactContextBaseJavaModule {
 		}
         map.putArray("recognizedLanguages", langToMap(line.getRecognizedLanguage()));
 
-        map.putDouble("confidenceScore", (double) line.getConfidence());
+        map.putDouble("confidenceScore", line.getConfidence());
 
-        map.putDouble("rotationDegree", (double) line.getAngle());
+        map.putDouble("rotationDegree", line.getAngle());
 
         WritableArray elements = Arguments.createArray();
         for (Text.Element element : line.getElements()) {
